Fix npc_text for classic.

+++ WowPacketParser/SQL/Builders/QuestMisc.cs
@@ -102,13 +102,13 @@ namespace WowPacketParser.SQL.Builders
                     requestItems.EmoteOnCompleteDelay = (uint)requestItemEmote.Value.EmoteOnCompleteDelay;
 
                 if (requestItemEmote.Value.EmoteOnComplete >= 0)
-                    requestItems.EmoteOnComplete = (uint)requestItemEmote.Value.EmoteOnComplete;
+                    requestItems.EmoteOnComplete = requestItemEmote.Value.EmoteOnComplete;
 
                 if (requestItemEmote.Value.EmoteOnIncompleteDelay >= 0)
                     requestItems.EmoteOnIncompleteDelay = (uint)requestItemEmote.Value.EmoteOnIncompleteDelay;
 
                 if (requestItemEmote.Value.EmoteOnIncomplete >= 0)
-                    requestItems.EmoteOnIncomplete = (uint)requestItemEmote.Value.EmoteOnIncomplete;
+                    requestItems.EmoteOnIncomplete = requestItemEmote.Value.EmoteOnIncomplete;
 
                 if (requestItemEmote.Value.EmoteOnCompleteDelay >= 0 && requestItemEmote.Value.EmoteOnComplete >= 0 && requestItemEmote.Value.EmoteOnIncompleteDelay >= 0 && requestItemEmote.Value.EmoteOnIncomplete >= 0)
                     requestItems.VerifiedBuild = ClientVersion.BuildInt;
